Move rail objects and prompt rails to where they're used
